06_Projects/copythatpay_system/backend_requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
ccxt>=4.0.0
python-dotenv>=1.0.0
aiohttp>=3.8.0
pandas>=2.0.0
numpy>=1.24.0
groq>=0.4.0
rich>=13.0.0
